refactor: rename to utils
